Moved NetworkCodec implementations to the API package.
