+++ library/core/networking/src/main/java/org/quiltmc/qsl/networking/api/codec/SimpleNetworkCodec.java
@@ -14,33 +14,43 @@
  * limitations under the License.
  */
 
-package org.quiltmc.qsl.networking.impl.codec;
+package org.quiltmc.qsl.networking.api.codec;
 
 import net.minecraft.network.PacketByteBuf;
 
 import org.quiltmc.qsl.networking.api.codec.NetworkCodec;
 
-public class NamedNetworkCodec<A> implements NetworkCodec<A> {
-	private final String name;
-	private final NetworkCodec<A> delegate;
+public final class SimpleNetworkCodec<A> implements NetworkCodec<A> {
+	private final PacketByteBuf.Reader<A> reader;
+	private final PacketByteBuf.Writer<A> writer;
 
-	public NamedNetworkCodec(String name, NetworkCodec<A> delegate) {
-		this.name = name;
-		this.delegate = delegate;
+	public SimpleNetworkCodec(PacketByteBuf.Writer<A> writer, PacketByteBuf.Reader<A> reader) {
+		this.reader = reader;
+		this.writer = writer;
 	}
 
 	@Override
 	public A decode(PacketByteBuf buf) {
-		return this.delegate.decode(buf);
+		return this.reader.apply(buf);
 	}
 
 	@Override
 	public void encode(PacketByteBuf buf, A data) {
-		this.delegate.encode(buf, data);
+		this.writer.accept(buf, data);
+	}
+
+	@Override
+	public PacketByteBuf.Reader<A> asReader() {
+		return this.reader;
+	}
+
+	@Override
+	public PacketByteBuf.Writer<A> asWriter() {
+		return this.writer;
 	}
 
 	@Override
 	public String toString() {
-		return this.name;
+		return "SimpleNetworkCodec";
 	}
 }
